Добавляет загрузку по 5 комментариев
