+++ js/full-size.js
@@ -1,34 +1,8 @@
+import {getComments} from './render-comments.js';
+
 const bigPicture = document.querySelector('.big-picture');
-const socialCommentsList = document.querySelector('.social__comments');
-const socialComment = document.querySelector('.social__comment');
-
-const renderComments = (comments) => {
-  socialCommentsList.innerHTML = '';
-  comments.forEach(({avatar, message}) => {
-    const userComment = socialComment.cloneNode(true);
-    userComment.querySelector('.social__picture').src = avatar;
-    userComment.querySelector('.social__text').textContent = message;
-    socialCommentsList.append(userComment);
-  });
-};
-
-const commentsCounter = document.querySelector('.social__comment-count');
-const commentsLoader = document.querySelector('.social__comments-loader');
-
-const createBigPicture = ({url, description, likes, comments}) => {
-  bigPicture.querySelector('.big-picture__img').querySelector('img').src = url;
-  bigPicture.querySelector('.social__caption').textContent = description;
-  bigPicture.querySelector('.likes-count').textContent = likes;
-  bigPicture.querySelector('.comments-count').textContent = comments.length;
-};
-
 const modalCloseElement = bigPicture.querySelector('.big-picture__cancel');
 const body = document.querySelector('body');
-
-function closeUserModal () {
-  bigPicture.classList.add('hidden');
-  body.classList.remove('modal-open');
-}
 
 const onEscapeKeydown = (evt) => {
   if (evt.key === 'Escape') {
@@ -37,16 +11,25 @@ const onEscapeKeydown = (evt) => {
   }
 };
 
+const createBigPicture = ({url, description, likes}) => {
+  bigPicture.querySelector('.big-picture__img').querySelector('img').src = url;
+  bigPicture.querySelector('.social__caption').textContent = description;
+  bigPicture.querySelector('.likes-count').textContent = likes;
+};
+
+function closeUserModal () {
+  bigPicture.classList.add('hidden');
+  body.classList.remove('modal-open');
+}
+
 modalCloseElement.addEventListener('click', () => closeUserModal());
+document.addEventListener('keydown', onEscapeKeydown);
 
 const getFullSize = (data) => {
   bigPicture.classList.remove('hidden');
   body.classList.add('modal-open');
-  commentsCounter.classList.add('hidden');
-  commentsLoader.classList.add('hidden');
-  document.addEventListener('keydown', onEscapeKeydown);
   createBigPicture(data);
-  renderComments(data.comments);
+  getComments(data);
 };
 
 export {getFullSize};
